Add -d command line option for direct processing
By the way of adding it, a bug in JPetOptionsTools::isProgressBar was
fixed: previously it read the value of the progress bar option with
map::at without checking if it exists, leading to an exception if the
option was not created before. We usually initialize the options to
default values and this did not see the problem before. Now, UT-s were
also added to check such cases.

Conflicts:
	include/Options/JPetOptionsTools/JPetOptionsTools.h
	src/Core/JPetCmdParser/JPetCmdParser.cpp
	src/Options/JPetOptionsGenerator/JPetOptionsGeneratorTools.cpp
	src/Options/JPetOptionsTools/JPetOptionsTools.cpp
	tests/Options/JPetOptionsTools/JPetOptionsToolsTest.cpp